# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exocortex Web API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/assets/{asset_name}/state": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get Asset State",
        "operationId": "get_asset_state_api_assets__asset_name__state_get",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStateModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/import": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Import Asset",
        "operationId": "import_asset_api_assets_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_asset_api_assets_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummaryModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/references/{name}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get Reference",
        "operationId": "get_reference_api_assets__asset_name__references__name__get",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/reveal": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Reveal Asset",
        "operationId": "reveal_asset_api_assets__asset_name__reveal_post",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/groups/{group_idx}/tutors/{tutor_idx}/questions": {
      "delete": {
        "tags": [
          "assets"
        ],
        "summary": "Delete Question",
        "operationId": "delete_question_api_assets__asset_name__groups__group_idx__tutors__tutor_idx__questions_delete",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          },
          {
            "name": "group_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Idx"
            }
          },
          {
            "name": "tutor_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tutor Idx"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/groups/{group_idx}/tutors": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Create Tutor Session",
        "operationId": "create_tutor_session_api_assets__asset_name__groups__group_idx__tutors_post",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          },
          {
            "name": "group_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Idx"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTutorRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorSessionModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}/groups/{group_idx}/tutors/{tutor_idx}": {
      "delete": {
        "tags": [
          "assets"
        ],
        "summary": "Delete Tutor Session",
        "operationId": "delete_tutor_session_api_assets__asset_name__groups__group_idx__tutors__tutor_idx__delete",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          },
          {
            "name": "group_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Idx"
            }
          },
          {
            "name": "tutor_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tutor Idx"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_name}": {
      "delete": {
        "tags": [
          "assets"
        ],
        "summary": "Delete Asset",
        "operationId": "delete_asset_api_assets__asset_name__delete",
        "parameters": [
          {
            "name": "asset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetStateAssetModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pageCount": {
            "type": "integer",
            "title": "Pagecount"
          },
          "pdfPath": {
            "type": "string",
            "title": "Pdfpath"
          }
        },
        "type": "object",
        "required": [
          "name",
          "pageCount",
          "pdfPath"
        ],
        "title": "AssetStateAssetModel"
      },
      "AssetStateModel": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetStateAssetModel"
          },
          "references": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "References"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/BlockModel"
            },
            "type": "array",
            "title": "Blocks"
          },
          "mergeOrder": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Mergeorder"
          },
          "nextBlockId": {
            "type": "integer",
            "title": "Nextblockid",
            "default": 1
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupModel"
            },
            "type": "array",
            "title": "Groups"
          },
          "uiState": {
            "$ref": "#/components/schemas/UiStateModel"
          }
        },
        "type": "object",
        "required": [
          "asset"
        ],
        "title": "AssetStateModel"
      },
...
      "BlockModel": {
        "properties": {
          "blockId": {
            "type": "integer",
            "title": "Blockid"
          },
          "pageIndex": {
            "type": "integer",
            "title": "Pageindex"
          },
          "fractionRect": {
            "$ref": "#/components/schemas/RectModel"
          },
          "groupIdx": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupidx"
          }
        },
        "type": "object",
        "required": [
          "blockId",
          "pageIndex",
          "fractionRect"
        ],
        "title": "BlockModel"
      },
      "Body_import_asset_api_assets_import_post": {
        "properties": {
          "source_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Source File"
          },
          "markdown_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Markdown File"
          },
          "content_list_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Content List File"
          },
          "asset_name": {
            "type": "string",
            "title": "Asset Name"
          },
          "asset_subfolder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Subfolder"
          }
        },
        "type": "object",
        "required": [
          "source_file",
          "markdown_file",
          "content_list_file",
          "asset_name"
        ],
        "title": "Body_import_asset_api_assets_import_post"
      },
...
      "CreateTutorRequest": {
        "properties": {
          "focusMarkdown": {
            "type": "string",
            "title": "Focusmarkdown"
          }
        },
        "type": "object",
        "required": [
          "focusMarkdown"
        ],
        "title": "CreateTutorRequest"
      },
...
      "GroupModel": {
        "properties": {
          "groupIdx": {
            "type": "integer",
            "title": "Groupidx"
          },
          "blockIds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Blockids"
          }
        },
        "type": "object",
        "required": [
          "groupIdx"
        ],
        "title": "GroupModel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "RectModel": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "height": {
            "type": "number",
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "RectModel"
      },
...
      "TaskSummaryModel": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "assetName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assetname"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "title",
          "createdAt",
          "updatedAt"
        ],
        "title": "TaskSummaryModel"
      },
      "TutorSessionModel": {
        "properties": {
          "tutorIdx": {
            "type": "integer",
            "title": "Tutoridx"
          },
          "markdownPath": {
            "type": "string",
            "title": "Markdownpath"
          }
        },
        "type": "object",
        "required": [
          "tutorIdx",
          "markdownPath"
        ],
        "title": "TutorSessionModel"
      },
...
      "UiStateModel": {
        "properties": {
          "currentPage": {
            "type": "integer",
            "title": "Currentpage",
            "default": 1
          },
          "zoom": {
            "type": "number",
            "title": "Zoom",
            "default": 1.0
          },
          "pdfScrollFraction": {
            "type": "number",
            "title": "Pdfscrollfraction",
            "default": 0.0
          },
          "pdfScrollLeftFraction": {
            "type": "number",
            "title": "Pdfscrollleftfraction",
            "default": 0.0
          },
          "currentMarkdownPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentmarkdownpath"
          },
          "openMarkdownPaths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Openmarkdownpaths"
          },
          "sidebarCollapsed": {
            "type": "boolean",
            "title": "Sidebarcollapsed",
            "default": false
          },
          "sidebarCollapsedNodeIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sidebarcollapsednodeids"
          },
          "markdownScrollFractions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Markdownscrollfractions"
          },
          "sidebarWidthRatio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sidebarwidthratio"
          },
          "rightRailWidthRatio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rightrailwidthratio"
          }
        },
        "type": "object",
        "title": "UiStateModel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}